requirements.txt
fastapi
uvicorn
pydantic
python-multipart
openai
google-generativeai
httpx
python-dotenv
